refactor(settings): apply better system design to code
